requirements.txt
configparser==3.5.0
confparse==1.0a1
dj-static==0.0.6
Django==1.9.1
django-environ==0.4.0
django-mptt==0.8.0
django-secure==1.0.1
django-widget-tweaks==1.4.1
gunicorn==19.6.0
mysqlclient
mistune==0.7.1
six==1.10.0
static3==0.7.0
wheel==0.24.0
dnspython==1.14.0
